+++ web/src/messages/SearchTypes.ml
@@ -1,0 +1,28 @@
+[@@@ocaml.warning "-27-30-39"]
+
+
+type search_suggestions_request = {
+  index : string;
+}
+
+type search_suggestions_response = {
+  task_types : string list;
+  authors : string list;
+  approvals : string list;
+}
+
+let rec default_search_suggestions_request 
+  ?index:((index:string) = "")
+  () : search_suggestions_request  = {
+  index;
+}
+
+let rec default_search_suggestions_response 
+  ?task_types:((task_types:string list) = [])
+  ?authors:((authors:string list) = [])
+  ?approvals:((approvals:string list) = [])
+  () : search_suggestions_response  = {
+  task_types;
+  authors;
+  approvals;
+}