+++ web/src/messages/TaskDataTypes.ml
@@ -5,7 +5,7 @@ type task_data_commit_request = {
   index : string;
   crawler : string;
   apikey : string;
-  timestamp : string;
+  timestamp : TimestampTypes.timestamp option;
 }
 
 type task_data_commit_error =
@@ -16,13 +16,26 @@ type task_data_commit_error =
 
 type task_data_commit_response =
   | Error of task_data_commit_error
-  | Timestamp of string
+  | Timestamp of TimestampTypes.timestamp
+
+type task_data_get_last_updated_error =
+  | Get_unknown_index 
+  | Get_unknown_crawler 
+
+type task_data_get_last_updated_request = {
+  index : string;
+  crawler : string;
+}
+
+type task_data_get_last_updated_response =
+  | Error of task_data_get_last_updated_error
+  | Timestamp of TimestampTypes.timestamp
 
 let rec default_task_data_commit_request 
   ?index:((index:string) = "")
   ?crawler:((crawler:string) = "")
   ?apikey:((apikey:string) = "")
-  ?timestamp:((timestamp:string) = "")
+  ?timestamp:((timestamp:TimestampTypes.timestamp option) = None)
   () : task_data_commit_request  = {
   index;
   crawler;
@@ -33,3 +46,15 @@ let rec default_task_data_commit_request
 let rec default_task_data_commit_error () = (Unknown_index:task_data_commit_error)
 
 let rec default_task_data_commit_response () : task_data_commit_response = Error (default_task_data_commit_error ())
+
+let rec default_task_data_get_last_updated_error () = (Get_unknown_index:task_data_get_last_updated_error)
+
+let rec default_task_data_get_last_updated_request 
+  ?index:((index:string) = "")
+  ?crawler:((crawler:string) = "")
+  () : task_data_get_last_updated_request  = {
+  index;
+  crawler;
+}
+
+let rec default_task_data_get_last_updated_response () : task_data_get_last_updated_response = Error (default_task_data_get_last_updated_error ())